Implement event manager system

+++ src/materials/dielectric.cpp
@@ -4,6 +4,7 @@
 #include "../types.h"
 
 #include <algorithm>
+#include <cmath>
 
 #include <glm/glm.hpp>
 
@@ -67,7 +68,7 @@ float Dielectric::reflectance(float cosTheta, float eta)
     // https://en.wikipedia.org/wiki/Schlick%27s_approximation
     float r0 = (1.0f - eta) / (1.0f + eta);
     r0 *= r0;
-    return r0 + (1.0f - r0) * powf(1.0f - cosTheta, 5);
+    return r0 + (1.0f - r0) * std::pow(1.0f - cosTheta, 5);
 }
